requirements.txt
viam-sdk==0.12.0
diskcache==5.6.3